move constructors for test *Loggers in the logger package
