+++ core/logger/test_logger.go
@@ -1,4 +1,4 @@
-package cltest
+package logger
 
 // Based on https://stackoverflow.com/a/52737940
 
@@ -7,8 +7,6 @@ import (
 	"log"
 	"net/url"
 	"sync"
-
-	"github.com/smartcontractkit/chainlink/core/logger"
 
 	"github.com/fatih/color"
 	"go.uber.org/zap"
@@ -48,7 +46,7 @@ var createSinkOnce sync.Once
 func registerMemorySink() {
 	testMemoryLog = &MemorySink{m: sync.Mutex{}, b: bytes.Buffer{}}
 	if err := zap.RegisterSink("memory", func(*url.URL) (zap.Sink, error) {
-		return logger.PrettyConsole{Sink: testMemoryLog}, nil
+		return PrettyConsole{Sink: testMemoryLog}, nil
 	}); err != nil {
 		panic(err)
 	}
@@ -61,7 +59,7 @@ func MemoryLogTestingOnly() *MemorySink {
 
 // CreateTestLogger creates a logger that directs output to PrettyConsole
 // configured for test output, and to the buffer testMemoryLog.
-func CreateTestLogger(lvl zapcore.Level) *zap.Logger {
+func CreateTestLogger(lvl zapcore.Level) *Logger {
 	_ = MemoryLogTestingOnly() // Make sure memory log is created
 	color.NoColor = false
 	config := zap.NewProductionConfig()
@@ -71,12 +69,14 @@ func CreateTestLogger(lvl zapcore.Level) *zap.Logger {
 	if err != nil {
 		log.Fatal(err)
 	}
-	return zl
+	return &Logger{
+		SugaredLogger: zl.Sugar(),
+	}
 }
 
 // CreateMemoryTestLogger creates a logger that only directs output to the
 // buffer testMemoryLog.
-func CreateMemoryTestLogger(lvl zapcore.Level) *zap.Logger {
+func CreateMemoryTestLogger(lvl zapcore.Level) *Logger {
 	_ = MemoryLogTestingOnly() // Make sure memory log is created
 	color.NoColor = true
 	config := zap.NewProductionConfig()
@@ -86,5 +86,7 @@ func CreateMemoryTestLogger(lvl zapcore.Level) *zap.Logger {
 	if err != nil {
 		log.Fatal(err)
 	}
-	return zl
+	return &Logger{
+		SugaredLogger: zl.Sugar(),
+	}
 }
